% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighCharts_Themes.R
\docType{data}
\name{color}
\alias{color}
\title{hc_theme_crs}
\format{An object of class \code{character} of length 8.}
\usage{
hc_theme_crs("crs.1")
}
\arguments{
\item{Currently, }{there is only one crs theme for use with highcharter package, 'crs.1'.}
}
\value{
highcharter font and color theme
}
\description{
This applies various crs color and font palette themes
}
\keyword{datasets}

