% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_to_existing_col.R
\name{move_to_existing_col}
\alias{move_to_existing_col}
\title{Move to Existing Column}
\usage{
move_to_existing_col(df, my.replacement.col, my.replacer.col, my.regex)
}
\arguments{
\item{my.replacement.col}{Column to which value will be moved}

\item{my.replacer.col}{Name of column that contain values, leaving behind NA}

\item{my.regex}{Regular expression}

\item{df}{dataframe}
}
\description{
Use regex to identify values move between columns, leaving behind NA
}
\details{
Wrapper for str_detect
}
\examples{
 
z <- tibble(col1 = letters[1:5], col2 = c("boo","bah", "beh", "bi", "E") )
move_to_newcol(z, col1, col2, "[A-Z]")


}
