% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_to_newcol.R
\name{move_to_newcol}
\alias{move_to_newcol}
\title{Move to New Column}
\usage{
move_to_newcol(df, col1, col2, my.regex)
}
\arguments{
\item{my.old.col.name}{Column from which value will be moved leaving behind NA}

\item{my.new.col.name}{Name of new column that will contain values.}

\item{my.regex}{Regular expression}

\item{df}{dataframe}
}
\description{
Use regex to identify values and move them to a new column, replacing original with NA
}
\details{
Wrapper for str_detect
}
\examples{
 
x <- tibble(col1 = letters[1:5])
move_to_newcol(x, col1, col2, "[a]")


}
