% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_Formats.R
\name{nettle_big.numbers}
\alias{nettle_big.numbers}
\title{nettle_big.numbers}
\usage{
nettle_big.numbers(num)
}
\arguments{
\item{num}{number to evaluate}
}
\description{
A function that abreviates big numbers and adds K (thousands), M (Millions), B (Billions), T (Trillions)
}
\examples{
big.numbers <- c(100000000, 98875, 86455, 2999999999)
nettle_big.numbers(big.numbers)




}

