% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettle_downzip.R
\name{nettle_downzip}
\alias{nettle_downzip}
\title{Download and Unzip}
\usage{
nettle_downzip(zip.url, zip.file)
}
\arguments{
\item{zip.url}{zip url}

\item{zip.file}{name of file inside zip that must be read}

\item{sheet}{read_excel}
}
\value{
location of unzipped file

 @examples x <- "http://comptroller.defense.gov/Portals/45/Documents/defbudget/fy2018/FY_2018_Green_Book.zip"
          y <- "FY18 PB Green Book Chap 2.xlsx" 
          my.path <- nettle_downzip(x, y)
          data <- read_excel(my.path, sheet = 1)
}
\description{
Download and Unzip file to temporary directory
}
\details{
This function will download and unzip file to temporary location, then
read_excl.
}
