% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nettle_export.R
\name{nettle_export}
\alias{nettle_export}
\title{nettle_export}
\usage{
nettle_export(df, my.filename, my.data.subfolder = "Processed")
}
\arguments{
\item{df}{Choose dataframe to export}

\item{my.filename}{Desired filename}

\item{my.data.subfolder}{Data/Raw' or Data/Processed' (default) subfolder}
}
\description{
This saves a timestamped file to declared subfolder using Hadley's write_csv function.
Intended for use with dataframes. Saves to either Data/Raw or Data/Processed (default) subfolders,
}

