% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiscal_Years.R
\name{nettle_fy}
\alias{nettle_fy}
\title{Define Fiscal Years}
\usage{
nettle_fy(x, firstMonth = 10, fy.prefix = "FY", quarter.prefix = "Q",
  sep = "-", level.range = c(min(x), max(x)))
}
\arguments{
\item{x}{vector of type \code{\link{Date}}.}

\item{firstMonth}{the month corresponding to the first month of the fiscal year (default is October)
Setting \code{firstMonth=1} is equivalent calendar years.}

\item{fy.prefix}{the character string to paste before the year.}

\item{quarter.prefix}{the character string to paste before the quarter.}

\item{sep}{the separater between the year and quarter.}

\item{level.range}{the range to use for defining the levels in the returned
factor.}
}
\description{
Return the year (fiscal or calendar) and quarter in which the date appears.
}
\details{
This function will cut the given date vector into quarters (i.e. three month
increments) and return an ordered factor with levels defined to be the quarters
between the minimum and maximum dates in the given vector. The levels, by
default, will be formated as \code{FY2013-Q1}, however the \code{FY} and \code{Q}
can be changed using the \code{fy.prefix} and \code{quarter.prefix} parameters,
respectively.

Original Script can be found at jbryer's Gist site:
https://gist.github.com/jbryer/5412193

Here's how to call the function from there directly:
 gistfile <-  "https://gist.github.com/jbryer/5412193"
 library(devtools)
 source_gist(gistfile)

The original function has been modified for Federal Calendar
(October 1 begins FY of following calendar year)
}
\examples{
dates <- as.Date(c('2013-04-03','2012-03-30','2011-10-31',
	                   '2011-04-14','2010-04-22','2004-10-04',
	                   '2000-02-29','1997-12-05','1997-04-23',
	                   '1997-04-01'))
	nettle_fy(dates)
	nettle_fydates, firstMonth=1)
	nettle_fy(dates, 1, '', '', '')
	\dontrun{
	nettle_fy(dates, level.range=as.Date(c('2010-01-01','2013-01-01')))
	}
}

