% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_Formats.R
\name{nettle_prettyNum}
\alias{nettle_prettyNum}
\title{nettle_prettyNum}
\usage{
nettle_prettyNum(c(1000, -1000, 1000000, -1000000, 1000000000, -1000000000, 1000000000000, -1000000000000))
}
\arguments{
\item{num}{A big or small number}
}
\description{
A function that formats big (and small) numbers
}
\details{
Numbers that have 12 digits are abreviated "T" for trillion
Numbers that have 9 digits are abreviated "B" for billion
Numbers that have 6 digits are abreviated "M" for Million
Commas are inserted. Some warning messages remain. Not sure why.
}
