% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.uniform.R
\name{nettle_uniform}
\alias{nettle_uniform}
\title{Uniformly format text}
\usage{
nettle_uniform(df$my.data)
}
\arguments{
\item{my.data}{List of data. Could be column, names(data), etc.}
}
\description{
Make syntactcially valid names out of character vectors
}
\details{
Useful for stripping out whitespace and formatting text according to pre- established R rules.
Converts characters to R-friendly, all caps. The character 'X' is prepended for
data that appears to be numeric. Invalid characters are translated to '.' Missing values
become 'NA.' For more details, refer to make.names function{base}
}

